Revert "Add `evm-tracing` feature (#778)"
This reverts commit 109582a7617ca88bbcd44b15d997b990c701f951.
