+++ node/cli/Cargo.toml
@@ -45,7 +45,6 @@ moonbeam-native = ["service/moonbeam-native"]
 
 wasmtime = ["sc-cli/wasmtime"]
 runtime-benchmarks = ["service/runtime-benchmarks"]
-evm-tracing = []
 try-runtime = [
 	"try-runtime-cli",
 	"service/try-runtime",
