+++ node/service/Cargo.toml
@@ -36,11 +36,10 @@ moonbase-runtime = { path = "../../runtime/moonbase", optional = true }
 
 moonbeam-rpc-txpool = { path = "../../client/rpc/txpool" }
 moonbeam-rpc-primitives-txpool = { path = "../../primitives/rpc/txpool" }
+moonbeam-rpc-debug = { path = "../../client/rpc/debug" }
 moonbeam-rpc-primitives-debug = { path = "../../primitives/rpc/debug" }
+moonbeam-rpc-trace = { path = "../../client/rpc/trace" }
 moonbeam-primitives-ext = { path = "../../primitives/ext" }
-
-moonbeam-rpc-debug = { path = "../../client/rpc/debug", optional = true }
-moonbeam-rpc-trace = { path = "../../client/rpc/trace", optional = true }
 
 moonbeam-core-primitives = { path = "../../core-primitives" }
 
@@ -148,14 +147,6 @@ default = [
 	"moonbase-native",
 	"moonriver-native",
 	"moonbeam-native"
-]
-
-evm-tracing = [
-	"moonbeam-rpc-debug",
-	"moonbeam-rpc-trace",
-	"moonbeam-runtime/evm-tracing",
-	"moonriver-runtime/evm-tracing",
-	"moonbase-runtime/evm-tracing"
 ]
 
 rococo-native = [ "polkadot-service/rococo-native", "polkadot-cli/rococo-native" ]
