+++ node/service/src/lib.rs
@@ -486,17 +486,7 @@ where
 	let subscription_task_executor =
 		sc_rpc::SubscriptionTaskExecutor::new(task_manager.spawn_handle());
 
-	rpc::spawn_essential_tasks(rpc::SpawnTasksParams {
-		task_manager: &task_manager,
-		client: client.clone(),
-		substrate_backend: backend.clone(),
-		frontier_backend: frontier_backend.clone(),
-		pending_transactions: pending_transactions.clone(),
-		filter_pool: filter_pool.clone(),
-	});
-
-	#[cfg(feature = "evm-tracing")]
-	let tracing_requesters = rpc::tracing::spawn_tracing_tasks(
+	let spawned_requesters = rpc::spawn_tasks(
 		&rpc_config,
 		rpc::SpawnTasksParams {
 			task_manager: &task_manager,
@@ -544,19 +534,14 @@ where
 				command_sink: None,
 				frontier_backend: frontier_backend.clone(),
 				backend: backend.clone(),
+				debug_requester: spawned_requesters.debug.clone(),
+				trace_filter_requester: spawned_requesters.trace.clone(),
+				trace_filter_max_count: rpc_config.ethapi_trace_max_count,
 				max_past_logs,
 				transaction_converter,
 			};
-			#[allow(unused_mut)]
-			let mut io = rpc::create_full(deps, subscription_task_executor.clone());
-			#[cfg(feature = "evm-tracing")]
-			rpc::tracing::extend_with_tracing(
-				client.clone(),
-				tracing_requesters.clone(),
-				rpc_config.ethapi_trace_max_count,
-				&mut io,
-			);
-			Ok(io)
+
+			Ok(rpc::create_full(deps, subscription_task_executor.clone()))
 		})
 	};
 
@@ -830,17 +815,7 @@ pub fn new_dev(
 		);
 	}
 
-	rpc::spawn_essential_tasks(rpc::SpawnTasksParams {
-		task_manager: &task_manager,
-		client: client.clone(),
-		substrate_backend: backend.clone(),
-		frontier_backend: frontier_backend.clone(),
-		pending_transactions: pending_transactions.clone(),
-		filter_pool: filter_pool.clone(),
-	});
-
-	#[cfg(feature = "evm-tracing")]
-	let tracing_requesters = rpc::tracing::spawn_tracing_tasks(
+	let spawned_requesters = rpc::spawn_tasks(
 		&rpc_config,
 		rpc::SpawnTasksParams {
 			task_manager: &task_manager,
@@ -886,19 +861,13 @@ pub fn new_dev(
 				command_sink: command_sink.clone(),
 				frontier_backend: frontier_backend.clone(),
 				backend: backend.clone(),
+				debug_requester: spawned_requesters.debug.clone(),
+				trace_filter_requester: spawned_requesters.trace.clone(),
+				trace_filter_max_count: rpc_config.ethapi_trace_max_count,
 				max_past_logs,
 				transaction_converter,
 			};
-			#[allow(unused_mut)]
-			let mut io = rpc::create_full(deps, subscription_task_executor.clone());
-			#[cfg(feature = "evm-tracing")]
-			rpc::tracing::extend_with_tracing(
-				client.clone(),
-				tracing_requesters.clone(),
-				rpc_config.ethapi_trace_max_count,
-				&mut io,
-			);
-			Ok(io)
+			Ok(rpc::create_full(deps, subscription_task_executor.clone()))
 		})
 	};
