+++ node/service/src/rpc.rs
@@ -16,16 +16,14 @@
 
 //! A collection of node-specific RPC extensions and related background tasks.
 
-#[cfg(feature = "evm-tracing")]
-pub mod tracing;
-
 use std::{sync::Arc, time::Duration};
 
 use fp_rpc::EthereumRuntimeRPCApi;
+use moonbeam_rpc_primitives_debug::DebugRuntimeApi;
 use sp_block_builder::BlockBuilder;
 
 use crate::{client::RuntimeApiCollection, TransactionConverters};
-use cli_opt::EthApi as EthApiCmd;
+use cli_opt::{EthApi as EthApiCmd, RpcConfig};
 use ethereum::EthereumStorageSchema;
 use fc_mapping_sync::{MappingSyncWorker, SyncStrategy};
 use fc_rpc::{
@@ -37,6 +35,11 @@ use fc_rpc_core::types::{FilterPool, PendingTransactions};
 use futures::StreamExt;
 use jsonrpc_pubsub::manager::SubscriptionManager;
 use moonbeam_core_primitives::{Block, Hash};
+use moonbeam_rpc_debug::DebugHandler;
+use moonbeam_rpc_debug::{Debug, DebugRequester, DebugServer};
+use moonbeam_rpc_trace::{
+	CacheRequester as TraceFilterCacheRequester, CacheTask, Trace, TraceServer,
+};
 use moonbeam_rpc_txpool::{TxPool, TxPoolServer};
 use pallet_transaction_payment_rpc::{TransactionPayment, TransactionPaymentApi};
 use sc_client_api::{
@@ -61,6 +64,12 @@ use sp_core::H256;
 use sp_runtime::traits::{BlakeTwo256, Block as BlockT};
 use std::collections::BTreeMap;
 use substrate_frame_rpc_system::{FullSystem, SystemApi};
+use tokio::sync::Semaphore;
+
+pub struct RpcRequesters {
+	pub debug: Option<DebugRequester>,
+	pub trace: Option<TraceFilterCacheRequester>,
+}
 
 /// Full client dependencies.
 pub struct FullDeps<C, P, A: ChainApi, BE> {
@@ -88,6 +97,12 @@ pub struct FullDeps<C, P, A: ChainApi, BE> {
 	pub backend: Arc<BE>,
 	/// Manual seal command sink
 	pub command_sink: Option<futures::channel::mpsc::Sender<EngineCommand<Hash>>>,
+	/// Debug server requester.
+	pub debug_requester: Option<DebugRequester>,
+	/// Trace filter cache server requester.
+	pub trace_filter_requester: Option<TraceFilterCacheRequester>,
+	/// Trace filter max count.
+	pub trace_filter_max_count: u32,
 	/// Maximum number of logs in a query.
 	pub max_past_logs: u32,
 	/// Ethereum transaction to Extrinsic converter.
@@ -124,6 +139,9 @@ where
 		command_sink,
 		frontier_backend,
 		backend: _,
+		debug_requester,
+		trace_filter_requester,
+		trace_filter_max_count,
 		max_past_logs,
 		transaction_converter,
 	} = deps;
@@ -206,6 +224,18 @@ where
 		);
 	};
 
+	if let Some(trace_filter_requester) = trace_filter_requester {
+		io.extend_with(TraceServer::to_delegate(Trace::new(
+			client,
+			trace_filter_requester,
+			trace_filter_max_count,
+		)));
+	}
+
+	if let Some(debug_requester) = debug_requester {
+		io.extend_with(DebugServer::to_delegate(Debug::new(debug_requester)));
+	}
+
 	io
 }
 
@@ -219,19 +249,49 @@ pub struct SpawnTasksParams<'a, B: BlockT, C, BE> {
 }
 
 /// Spawn the tasks that are required to run Moonbeam.
-pub fn spawn_essential_tasks<B, C, BE>(params: SpawnTasksParams<B, C, BE>)
+pub fn spawn_tasks<B, C, BE>(
+	rpc_config: &RpcConfig,
+	params: SpawnTasksParams<B, C, BE>,
+) -> RpcRequesters
 where
 	C: ProvideRuntimeApi<B> + BlockOf,
 	C: HeaderBackend<B> + HeaderMetadata<B, Error = BlockChainError> + 'static,
 	C: BlockchainEvents<B>,
 	C: Send + Sync + 'static,
-	C::Api: EthereumRuntimeRPCApi<B>,
+	C::Api: EthereumRuntimeRPCApi<B> + DebugRuntimeApi<B> + DebugRuntimeApi<B>,
 	C::Api: BlockBuilder<B>,
 	B: BlockT<Hash = H256> + Send + Sync + 'static,
 	B::Header: HeaderT<Number = u32>,
 	BE: Backend<B> + 'static,
 	BE::State: StateBackend<BlakeTwo256>,
 {
+	let permit_pool = Arc::new(Semaphore::new(rpc_config.ethapi_max_permits as usize));
+
+	let (trace_filter_task, trace_filter_requester) =
+		if rpc_config.ethapi.contains(&EthApiCmd::Trace) {
+			let (trace_filter_task, trace_filter_requester) = CacheTask::create(
+				Arc::clone(&params.client),
+				Arc::clone(&params.substrate_backend),
+				Duration::from_secs(rpc_config.ethapi_trace_cache_duration),
+				Arc::clone(&permit_pool),
+			);
+			(Some(trace_filter_task), Some(trace_filter_requester))
+		} else {
+			(None, None)
+		};
+
+	let (debug_task, debug_requester) = if rpc_config.ethapi.contains(&EthApiCmd::Debug) {
+		let (debug_task, debug_requester) = DebugHandler::task(
+			Arc::clone(&params.client),
+			Arc::clone(&params.substrate_backend),
+			Arc::clone(&params.frontier_backend),
+			Arc::clone(&permit_pool),
+		);
+		(Some(debug_task), Some(debug_requester))
+	} else {
+		(None, None)
+	};
+
 	// Frontier offchain DB task. Essential.
 	// Maps emulated ethereum data to substrate native data.
 	params.task_manager.spawn_essential_handle().spawn(
@@ -246,6 +306,24 @@ where
 		)
 		.for_each(|()| futures::future::ready(())),
 	);
+
+	// `trace_filter` cache task. Essential.
+	// Proxies rpc requests to it's handler.
+	if let Some(trace_filter_task) = trace_filter_task {
+		params
+			.task_manager
+			.spawn_essential_handle()
+			.spawn("trace-filter-cache", trace_filter_task);
+	}
+
+	// `debug` task if enabled. Essential.
+	// Proxies rpc requests to it's handler.
+	if let Some(debug_task) = debug_task {
+		params
+			.task_manager
+			.spawn_essential_handle()
+			.spawn("ethapi-debug", debug_task);
+	}
 
 	// Frontier `EthFilterApi` maintenance.
 	// Manages the pool of user-created Filters.
@@ -283,4 +361,9 @@ where
 			Arc::clone(&params.frontier_backend),
 		),
 	);
+
+	RpcRequesters {
+		debug: debug_requester,
+		trace: trace_filter_requester,
+	}
 }
