+++ primitives/rpc/debug/Cargo.toml
@@ -23,14 +23,13 @@ sp-core = { git = "https://github.com/purestake/substrate.git", branch = "crysta
 serde = { version = "1.0", features = ["derive"], optional = true }
 serde_json = { version = "1.0", optional = true }
 hex = { version = "0.4", features = ["serde"], optional = true}
-evm = { version = "0.30.1", default-features = false, features = ["with-codec"] }
-evm-runtime = { version = "0.30.0", default-features = false }
-evm-gasometer = { version = "0.30.0", default-features = false }
+evm = { version = "0.30.1", default-features = false, features = ["with-codec", "tracing"] }
+evm-runtime = { version = "0.30.0", default-features = false, features = ["tracing"] }
+evm-gasometer = { version = "0.30.0", default-features = false, features = ["tracing"] }
 #moonbeam-rpc-core-debug = { path = "../../../client/rpc-core/debug" }
 
 [features]
 default = ["std"]
-evm-tracing = ["evm/tracing", "evm-runtime/tracing", "evm-gasometer/tracing"]
 std = [
 	"codec/std",
 	"sp-api/std",
