+++ primitives/rpc/debug/src/proxy/types/evm_gasometer_types.rs
@@ -67,7 +67,6 @@ pub enum GasometerEvent {
 	},
 }
 
-#[cfg(feature = "evm-tracing")]
 impl From<evm_gasometer::tracing::Event> for GasometerEvent {
 	fn from(i: evm_gasometer::tracing::Event) -> Self {
 		match i {
