+++ primitives/rpc/debug/src/proxy/types/evm_types.rs
@@ -150,7 +150,6 @@ pub enum EvmEvent {
 	},
 }
 
-#[cfg(feature = "evm-tracing")]
 impl<'a> From<evm::tracing::Event<'a>> for EvmEvent {
 	fn from(i: evm::tracing::Event<'a>) -> Self {
 		match i {
