+++ runtime/evm_tracer/Cargo.toml
@@ -11,9 +11,9 @@ repository = 'https://github.com/PureStake/moonbeam/'
 codec = { package = "parity-scale-codec", version = "2.2", default-features = false }
 fp-evm = { git = "https://github.com/purestake/frontier", default-features = false, branch = "notlesh-v0.9.9-block-response-length" }
 pallet-evm = { git = "https://github.com/purestake/frontier", default-features = false, branch = "notlesh-v0.9.9-block-response-length" }
-evm = { version = "0.30.1", default-features = false, features = ["with-codec"] }
-evm-runtime = { version = "0.30.0", default-features = false }
-evm-gasometer = { version = "0.30.0", default-features = false }
+evm = { version = "0.30.1", default-features = false, features = ["with-codec", "tracing"] }
+evm-runtime = { version = "0.30.0", default-features = false, features = ["tracing"] }
+evm-gasometer = { version = "0.30.0", default-features = false, features = ["tracing"] }
 sp-core = { default-features = false, git = "https://github.com/purestake/substrate.git", branch = "crystalin-v0.9.9-block-response-length" }
 sp-std = { default-features = false, git = "https://github.com/purestake/substrate.git", branch = "crystalin-v0.9.9-block-response-length" }
 sp-io = { default-features = false, git = "https://github.com/purestake/substrate.git", branch = "crystalin-v0.9.9-block-response-length" }
@@ -24,7 +24,6 @@ ethereum-types = { version = "0.12.0", default-features = false }
 
 [features]
 default = ["std"]
-evm-tracing = ["evm/tracing", "evm-runtime/tracing", "evm-gasometer/tracing"]
 std = [
 	"codec/std",
 	"evm/std",
