+++ tests/util/constants.ts
@@ -7,7 +7,6 @@ export const MOONBEAM_LOG = process.env.MOONBEAM_LOG || "info";
 export const BINARY_PATH = process.env.BINARY_PATH || `../target/release/moonbeam`;
 export const RELAY_BINARY_PATH = process.env.RELAY_BINARY_PATH || `../target/release/polkadot`;
 export const SPAWNING_TIME = 20000;
-export const ETHAPI_CMD = process.env.ETHAPI_CMD || "";
 
 // Test variables
 export const ZERO_ADDRESS = "0x0000000000000000000000000000000000000000";
